alignment_service/main.py

import io
import os
import base64
from typing import Optional, List, Dict, Any
from fastapi import FastAPI, File, UploadFile, Form
from fastapi.responses import JSONResponse
from pydantic import BaseModel
from striprtf.striprtf import rtf_to_text

# Optional: try to import faster-whisper; if not available, we fall back to a stub
USE_WHISPER = True
try:
    from faster_whisper import WhisperModel  # type: ignore
except Exception as e:
    USE_WHISPER = False

app = FastAPI(title="Case Scribe Alignment Service")

class ErrorItem(BaseModel):
    line: int
    column: int
    original: str
    suggested: str
    confidence: float
    type: str

class Summary(BaseModel):
    totalErrors: int
    byType: Dict[str, int]
    confidenceScore: float
    processingTime: float

class Analysis(BaseModel):
    errors: List[ErrorItem]
    summary: Summary

def _tokenize(text: str):
    import re
    tokens = re.findall(r"\w+|[^\w\s]", text, re.UNICODE)
    return tokens

def _align_stub(transcript_text: str, audio_text: str) -> List[ErrorItem]:
    # Very naive diff-based mismatch detector as placeholder.
    import difflib
    t_tokens = _tokenize(transcript_text.lower())
    a_tokens = _tokenize(audio_text.lower())
    sm = difflib.SequenceMatcher(a=t_tokens, b=a_tokens)
    errors: List[ErrorItem] = []
    line = 1
    col = 1
    for tag, i1, i2, j1, j2 in sm.get_opcodes():
        if tag == "replace" or tag == "delete":
            original = " ".join(t_tokens[i1:i2])[:80]
            suggested = " ".join(a_tokens[j1:j2])[:80]
            errors.append(ErrorItem(
                line=line,
                column=col,
                original=original or "(missing)",
                suggested=suggested or "(remove)",
                confidence=0.75,
                type="audio_mismatch"
            ))
            col += 1
        elif tag == "insert":
            original = ""
            suggested = " ".join(a_tokens[j1:j2])[:80]
            errors.append(ErrorItem(
                line=line,
                column=col,
                original="(missing)",
                suggested=suggested,
                confidence=0.7,
                type="audio_mismatch"
            ))
            col += 1
    return errors

def _grammar_spell_flags(transcript_text: str) -> List[ErrorItem]:
    # Placeholder for grammar/spelling; in production use a model or libraries like language_tool_python
    suspicious = []
    for bad, good in [("councelor", "counselor"), ("inadmissable", "inadmissible")]:
        if bad in transcript_text:
            suspicious.append(ErrorItem(
                line=1, column=1, original=bad, suggested=good, confidence=0.95, type="spelling"
            ))
    return suspicious

def build_docx(corrected_text: str) -> bytes:
    from docx import Document
    from docx.shared import Pt
    doc = Document()
    style = doc.styles['Normal']
    font = style.font
    font.name = 'Calibri'
    font.size = Pt(11)
    for para in corrected_text.split("\n\n"):
        doc.add_paragraph(para)
    bio = io.BytesIO()
    doc.save(bio)
    return bio.getvalue()

@app.get("/health")
def health():
    return {"ok": True, "use_whisper": USE_WHISPER}

@app.post("/align")
async def align_endpoint(
    rtf_file: Optional[UploadFile] = File(default=None),
    audio_file: UploadFile = File(...),
    plain_text: Optional[str] = Form(default=None),
    whisper_model_size: str = Form(default="small")
):
    # 1) Load transcript text (prefer RTF -> text; else plain text)
    transcript_text = ""
    if rtf_file is not None:
        raw = await rtf_file.read()
        try:
            transcript_text = rtf_to_text(raw.decode('utf-8', errors='ignore'))
        except Exception:
            # fallback treat it as plain text
            transcript_text = raw.decode('utf-8', errors='ignore')
    elif plain_text:
        transcript_text = plain_text
    else:
        return JSONResponse(status_code=400, content={"error": "No transcript provided (rtf_file or plain_text)."})
    transcript_text = transcript_text.strip()

    # 2) Transcribe audio (or stub)
    audio_bytes = await audio_file.read()

    audio_text = ""
    if USE_WHISPER:
        try:
            # load model (CPU by default; for GPU set envs)
            model = WhisperModel(whisper_model_size, device="cpu", compute_type="int8")
            import tempfile, os
            with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as tmp:
                tmp.write(audio_bytes)
                tmp_path = tmp.name
            segments, info = model.transcribe(tmp_path, word_timestamps=True)
            words = []
            for seg in segments:
                if hasattr(seg, "words") and seg.words:
                    for w in seg.words:
                        words.append(w.word)
                else:
                    words.append(seg.text)
            audio_text = " ".join(words).strip()
            try:
                os.remove(tmp_path)
            except Exception:
                pass
        except Exception as e:
            # fallback
            audio_text = ""
            print("Whisper failed:", e)
    if not audio_text:
        # fallback to treating audio as empty; alignment will produce many mismatches
        audio_text = ""

    # 3) Alignment + error flags
    errors = _align_stub(transcript_text, audio_text)
    errors += _grammar_spell_flags(transcript_text)

    # summary
    by_type = {}
    for e in errors:
        by_type[e.type] = by_type.get(e.type, 0) + 1
    summary = Summary(
        totalErrors=len(errors),
        byType=by_type,
        confidenceScore=0.8 if errors else 0.95,
        processingTime=7.2
    )
    analysis = Analysis(errors=errors, summary=summary)

    # 4) Corrected transcript (for now, we just prefer transcript_text and apply trivial substitutions)
    corrected = transcript_text.replace("councelor", "counselor").replace("inadmissable", "inadmissible")

    # 5) Build downloadable files (txt, docx)
    txt_bytes = corrected.encode('utf-8')
    docx_bytes = build_docx(corrected)

    return {
        "analysis": analysis.model_dump(),
        "correctedTranscript": corrected,
        "downloads": {
            "txt_base64": base64.b64encode(txt_bytes).decode('ascii'),
            "docx_base64": base64.b64encode(docx_bytes).decode('ascii'),
            "filenames": {
                "txt": "corrected_transcript.txt",
                "docx": "corrected_transcript.docx"
            }
        }
    }
