alignment_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.2
python-docx==1.1.2
faster-whisper==1.0.3
striprtf==0.0.26
numpy==1.26.4
